Add Precedence information to an extended grammar
