+++ parser/lr/conflict.go
@@ -10,70 +10,6 @@ import (
 	"github.com/moorara/algo/sort"
 )
 
-var (
-	eqPrecedenceHandle = func(lhs, rhs *PrecedenceHandle) bool {
-		return lhs.Equal(rhs)
-	}
-)
-
-// PrecedenceHandle represents either a terminal symbol or a production rule
-// in the context of determining precedence for conflict resolution.
-type PrecedenceHandle struct {
-	*grammar.Terminal
-	*grammar.Production
-}
-
-// IsTerminal returns true if the handle represents a terminal symbol.
-func (h *PrecedenceHandle) IsTerminal() bool {
-	return h.Terminal != nil && h.Production == nil
-}
-
-// IsProduction returns true if the handle represents a production rule.
-func (h *PrecedenceHandle) IsProduction() bool {
-	return h.Terminal == nil && h.Production != nil
-}
-
-// String returns a string representation of the handle.
-func (h *PrecedenceHandle) String() string {
-	if h.IsTerminal() {
-		return h.Terminal.String()
-	} else if h.IsProduction() {
-		return fmt.Sprintf("%s = %s", h.Production.Head, h.Production.Body)
-	}
-
-	panic("PrecedenceHandle.String: invalid configuration")
-}
-
-// Equal determines whether or not two handles are the same.
-func (h *PrecedenceHandle) Equal(rhs *PrecedenceHandle) bool {
-	switch {
-	case h.IsTerminal() && rhs.IsTerminal():
-		return grammar.EqTerminal(*h.Terminal, *rhs.Terminal)
-	case h.IsProduction() && rhs.IsProduction():
-		return grammar.EqProduction(h.Production, rhs.Production)
-
-	}
-
-	return false
-}
-
-// cmpPrecedenceHandle compares two handles and establishes an order between them.
-// Terminal handles come before Production handles.
-func cmpPrecedenceHandle(lhs, rhs *PrecedenceHandle) int {
-	switch {
-	case lhs.IsTerminal() && rhs.IsProduction():
-		return -1
-	case lhs.IsProduction() && rhs.IsTerminal():
-		return 1
-	case lhs.IsTerminal() && rhs.IsTerminal():
-		return grammar.CmpTerminal(*lhs.Terminal, *rhs.Terminal)
-	case lhs.IsProduction() && rhs.IsProduction():
-		return grammar.CmpProduction(lhs.Production, rhs.Production)
-	}
-
-	panic("cmpPrecedenceHandle: invalid configuration")
-}
-
 // ConflictError represents a conflict in an LR parsing table.
 // A conflict occurs when the grammar is ambiguous, resulting in multiple actions
 // being associated with a specific state s and terminal a in the ACTION table.
