+++ parser/lr/conflict_test.go
@@ -9,169 +9,6 @@ import (
 	"github.com/moorara/algo/set"
 )
 
-func TestPrecedenceHandle(t *testing.T) {
-	id := grammar.Terminal("ID")
-	num := grammar.Terminal("NUM")
-
-	type EqualTest struct {
-		rhs           *PrecedenceHandle
-		expectedEqual bool
-	}
-
-	tests := []struct {
-		name                 string
-		h                    *PrecedenceHandle
-		expectedIsTerminal   bool
-		expectedIsProduction bool
-		expectedString       string
-		equalTests           []EqualTest
-	}{
-		{
-			name: "Terminal",
-			h: &PrecedenceHandle{
-				Terminal: &id,
-			},
-			expectedIsTerminal:   true,
-			expectedIsProduction: false,
-			expectedString:       `"ID"`,
-			equalTests: []EqualTest{
-				{
-					rhs: &PrecedenceHandle{
-						Terminal: &id,
-					},
-					expectedEqual: true,
-				},
-				{
-					rhs: &PrecedenceHandle{
-						Terminal: &num,
-					},
-					expectedEqual: false,
-				},
-				{
-					rhs: &PrecedenceHandle{
-						Production: &grammar.Production{
-							Head: "rhs",
-							Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.NonTerminal("rhs")},
-						},
-					},
-					expectedEqual: false,
-				},
-			},
-		},
-		{
-			name: "Production",
-			h: &PrecedenceHandle{
-				Production: &grammar.Production{
-					Head: "rhs",
-					Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.NonTerminal("rhs")},
-				},
-			},
-			expectedIsTerminal:   false,
-			expectedIsProduction: true,
-			expectedString:       `rhs = rhs rhs`,
-			equalTests: []EqualTest{
-				{
-					rhs: &PrecedenceHandle{
-						Production: &grammar.Production{
-							Head: "rhs",
-							Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.NonTerminal("rhs")},
-						},
-					},
-					expectedEqual: true,
-				},
-				{
-					rhs: &PrecedenceHandle{
-						Production: &grammar.Production{
-							Head: "rhs",
-							Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.Terminal("|"), grammar.NonTerminal("rhs")},
-						},
-					},
-					expectedEqual: false,
-				},
-				{
-					rhs: &PrecedenceHandle{
-						Terminal: &id,
-					},
-					expectedEqual: false,
-				},
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.expectedIsTerminal, tc.h.IsTerminal())
-			assert.Equal(t, tc.expectedIsProduction, tc.h.IsProduction())
-			assert.Equal(t, tc.expectedString, tc.h.String())
-
-			t.Run("Equal", func(t *testing.T) {
-				for _, test := range tc.equalTests {
-					assert.Equal(t, test.expectedEqual, tc.h.Equal(test.rhs))
-				}
-			})
-		})
-	}
-}
-
-func TestCmpPrecedenceHandle(t *testing.T) {
-	id := grammar.Terminal("ID")
-	num := grammar.Terminal("NUM")
-
-	tests := []struct {
-		name            string
-		lhs             *PrecedenceHandle
-		rhs             *PrecedenceHandle
-		expectedCompare int
-	}{
-		{
-			name: "BothTerminal",
-			lhs: &PrecedenceHandle{
-				Terminal: &id,
-			},
-			rhs: &PrecedenceHandle{
-				Terminal: &num,
-			},
-			expectedCompare: -1,
-		},
-		{
-			name: "BothProduction",
-			lhs: &PrecedenceHandle{
-				Production: &grammar.Production{
-					Head: "rhs",
-					Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.NonTerminal("rhs")},
-				},
-			},
-			rhs: &PrecedenceHandle{
-				Production: &grammar.Production{
-					Head: "rhs",
-					Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.Terminal("|"), grammar.NonTerminal("rhs")},
-				},
-			},
-			expectedCompare: 1,
-		},
-		{
-			name: "Mixed",
-			lhs: &PrecedenceHandle{
-				Terminal: &id,
-			},
-			rhs: &PrecedenceHandle{
-				Production: &grammar.Production{
-					Head: "rhs",
-					Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.NonTerminal("rhs")},
-				},
-			},
-			expectedCompare: -1,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.expectedCompare, cmpPrecedenceHandle(tc.lhs, tc.rhs))
-
-		})
-	}
-}
-
 func TestConflictError(t *testing.T) {
 	star := grammar.Terminal("*")
 	id := grammar.Terminal("id")
