+++ parser/lr/lookahead/parsing_table.go
@@ -25,20 +25,20 @@ func BuildParsingTable(G *grammar.CFG) (*lr.ParsingTable, error) {
 	 * OUTPUT: The LALR parsing table functions ACTION and GOTO for G′.
 	 */
 
-	auto1 := lr.NewLR1KernelAutomaton(G)
+	H := lr.NewGrammarWithLR1Kernel(G, nil)
 
 	K := ComputeLALR1Kernels(G) // 1. Construct the kernels of the LALR(1) collection of sets of items for G′.
 	S := lr.BuildStateMap(K)    // Map sets of LALR(1) items to state numbers.
 
-	terminals := auto1.G().OrderTerminals()
-	_, _, nonTerminals := auto1.G().OrderNonTerminals()
+	terminals := H.OrderTerminals()
+	_, _, nonTerminals := H.OrderNonTerminals()
 	table := lr.NewParsingTable(S.States(), terminals, nonTerminals)
 
 	// 2. State i is constructed from I.
 	for i, I := range S.All() {
 		// The parsing action for state i is determined as follows:
 
-		for item := range auto1.CLOSURE(I).All() {
+		for item := range H.CLOSURE(I).All() {
 			item := item.(*lr.Item1)
 
 			// If "A → α•aβ, b" is in Iᵢ and GOTO(Iᵢ,a) = Iⱼ (a must be a terminal)
@@ -50,7 +50,7 @@ func BuildParsingTable(G *grammar.CFG) (*lr.ParsingTable, error) {
 					// Let K be the union of all item sets that have the same core as GOTO(I₁, X). Then, GOTO(J, X) = K.
 					// This means that the regular GOTO function returns an item set whose core matches one of the LALR(1) item sets, but it may have missing lookaheads.
 					// Therefore, to find K, we need to identify an item set whose core is a superset of the set returned by the regular GOTO function.
-					J := auto1.GOTO(I, a)
+					J := H.GOTO(I, a)
 					j := findSuperset(S, J)
 
 					// Set ACTION[i,a] to SHIFT j
@@ -86,15 +86,15 @@ func BuildParsingTable(G *grammar.CFG) (*lr.ParsingTable, error) {
 
 		// 3. The goto transitions for state i are constructed for all non-terminals A using the rule:
 		// If GOTO(Iᵢ,A) = Iⱼ
-		for A := range auto1.G().NonTerminals.All() {
-			if !A.Equal(auto1.G().Start) {
+		for A := range H.NonTerminals.All() {
+			if !A.Equal(H.Start) {
 				// J is the union of multiple sets of LR(1) items (J = I₁ ∪ I₂ ∪ ... ∪ Iₖ).
 				// By definition, I₁, I₂, ..., Iₖ all share the same core, which consists of the production and dot positions, excluding the lookahead.
 				// The cores of GOTO(I₁, X), GOTO(I₂, X), ..., GOTO(Iₖ, X) are also the same because the sets I₁, I₂, ..., Iₖ have identical cores.
 				// Let K be the union of all item sets that have the same core as GOTO(I₁, X). Then, GOTO(J, X) = K.
 				// This means that the regular GOTO function returns an item set whose core matches one of the LALR(1) item sets, but it may have missing lookaheads.
 				// Therefore, to find K, we need to identify an item set whose core is a superset of the set returned by the regular GOTO function.
-				J := auto1.GOTO(I, A)
+				J := H.GOTO(I, A)
 				j := findSuperset(S, J)
 
 				// Set GOTO[i,A] = j
@@ -151,10 +151,10 @@ func ComputeLALR1Kernels(G *grammar.CFG) lr.ItemSetCollection {
 	 *      We continue making passes over the kernel items until no more new lookaheads are propagated.
 	 */
 
-	auto0 := lr.NewLR0KernelAutomaton(G)
-	auto1 := lr.NewLR1KernelAutomaton(G)
+	H0 := lr.NewGrammarWithLR0Kernel(G, nil)
+	H1 := lr.NewGrammarWithLR1Kernel(G, nil)
 
-	K0 := auto0.Canonical()    // Construct the kernels of the sets of LR(0) items for G′.
+	K0 := H0.Canonical()       // Construct the kernels of the sets of LR(0) items for G′.
 	S0 := lr.BuildStateMap(K0) // Map Kernel sets of LR(0) items to state numbers.
 
 	propagations := newPropagationTable(S0) // This table memoize which items propagate their lookaheads to which other items.
@@ -181,7 +181,7 @@ func ComputeLALR1Kernels(G *grammar.CFG) lr.ItemSetCollection {
 			from := &scopedItem{ItemSet: s, Item: i}
 
 			// J = CLOSURE({[A → α.β, $]})
-			J := auto1.CLOSURE(
+			J := H1.CLOSURE(
 				lr.NewItemSet(
 					item.(*lr.Item0).Item1(grammar.Endmarker),
 				),
@@ -189,8 +189,8 @@ func ComputeLALR1Kernels(G *grammar.CFG) lr.ItemSetCollection {
 
 			for j := range J.All() {
 				if X, ok := j.DotSymbol(); ok {
-					nextI := auto0.GOTO(I, X) // GOTO(I,X)
-					nextj, _ := j.Next()      // B → γX.δ
+					nextI := H0.GOTO(I, X) // GOTO(I,X)
+					nextj, _ := j.Next()   // B → γX.δ
 
 					// Find B → γX.δ in GOTO(I,X)
 					to := new(scopedItem)
