+++ parser/lr/simple/parsing_table.go
@@ -15,15 +15,15 @@ func BuildParsingTable(G *grammar.CFG) (*lr.ParsingTable, error) {
 	 * OUTPUT: The SLR parsing table functions ACTION and GOTO for G′.
 	 */
 
-	auto0 := lr.NewLR0Automaton(G)
-	FIRST := auto0.G().ComputeFIRST()
-	FOLLOW := auto0.G().ComputeFOLLOW(FIRST)
+	H := lr.NewGrammarWithLR0(G, nil)
+	FIRST := H.ComputeFIRST()
+	FOLLOW := H.ComputeFOLLOW(FIRST)
 
-	C := auto0.Canonical()   // 1. Construct C = {I₀, I₁, ..., Iₙ}, the collection of sets of LR(0) items for G′.
+	C := H.Canonical()       // 1. Construct C = {I₀, I₁, ..., Iₙ}, the collection of sets of LR(0) items for G′.
 	S := lr.BuildStateMap(C) // Map sets of LR(0) items to state numbers.
 
-	terminals := auto0.G().OrderTerminals()
-	_, _, nonTerminals := auto0.G().OrderNonTerminals()
+	terminals := H.OrderTerminals()
+	_, _, nonTerminals := H.OrderNonTerminals()
 	table := lr.NewParsingTable(S.States(), terminals, nonTerminals)
 
 	// 2. State i is constructed from I.
@@ -36,7 +36,7 @@ func BuildParsingTable(G *grammar.CFG) (*lr.ParsingTable, error) {
 			// If "A → α•aβ" is in Iᵢ and GOTO(Iᵢ,a) = Iⱼ (a must be a terminal)
 			if X, ok := item.DotSymbol(); ok {
 				if a, ok := X.(grammar.Terminal); ok {
-					J := auto0.GOTO(I, a)
+					J := H.GOTO(I, a)
 					j := S.FindItemSet(J)
 
 					// Set ACTION[i,a] to SHIFT j
@@ -83,9 +83,9 @@ func BuildParsingTable(G *grammar.CFG) (*lr.ParsingTable, error) {
 
 		// 3. The goto transitions for state i are constructed for all non-terminals A using the rule:
 		// If GOTO(Iᵢ,A) = Iⱼ
-		for A := range auto0.G().NonTerminals.All() {
-			if !A.Equal(auto0.G().Start) {
-				J := auto0.GOTO(I, A)
+		for A := range H.NonTerminals.All() {
+			if !A.Equal(H.Start) {
+				J := H.GOTO(I, A)
 				j := S.FindItemSet(J)
 
 				// Set GOTO[i,A] = j
